Split register email and verify otp into components
